sửa lại project

+++ controllers/orderController.js
@@ -12,15 +12,16 @@ async function getOrderByOrderID(req, res) {
 
 
         if (result.recordset.length === 0) {
-            return res.status(404).send("Không tìm thấy đơn hàng");
+            return res.status(404).json({ message: "Không tìm thấy đơn hàng" });
         
         }
         const order = result.recordset[0];
         
         if (req.user.Role !== 'Admin' && req.user.UserID !== order.UserID) {
-            return res.status(403).send("Bạn không có quyền truy cập đơn hàng này!");
+            return res.status(403).json({ message: "Bạn không có quyền truy cập đơn hàng này!" });
         }
         const resultUser = await sql.query`SELECT FullName, Email FROM Users WHERE UserID = ${order.UserID}`;
+        
         res.send({
             user: resultUser.recordset[0],
             orders: [order]
@@ -39,22 +40,41 @@ async function getAllOrders(req, res) {
                 FROM Users u
                 JOIN Orders o ON u.UserID = o.UserID
                 JOIN Products p ON o.ProductID = p.ProductID
-                `;
+            `;
+            if (result.recordset.length === 0) {
+                return res.status(403).json({ message: "Hiện không có đơn hàng nào!" });
+            }
+            return res.send(result.recordset);
+        } else {
+            const userID = req.user.UserID;
 
-            res.send(result.recordset);
-
-        }
-        else {
-            const userID = parseInt(req.user.UserID);
-            const result = await sql.query`
+            const ordersResult = await sql.query`
                 SELECT o.*, p.ProductName
                 FROM Orders o
                 JOIN Products p ON o.ProductID = p.ProductID
                 WHERE o.UserID = ${userID}
             `;
-            res.send(result.recordset);
-        } 
 
+            if (ordersResult.recordset.length === 0) {
+                return res.status(200).send({ message: "Bạn chưa có đơn hàng nào" });
+            }
+
+            const userInfoResult = await sql.query`
+                SELECT FullName, Email
+                FROM Users
+                WHERE UserID = ${userID}
+            `;
+
+            const userInfo = userInfoResult.recordset[0];
+
+            return res.send({
+                User: {
+                    FullName: userInfo.FullName,
+                    Email: userInfo.Email
+                },
+                Orders: ordersResult.recordset
+            });
+        }
     } catch (err) {
         res.status(500).send("Lỗi: " + err.message);
     }
@@ -62,21 +82,34 @@ async function getAllOrders(req, res) {
 
 async function addOrders(req, res) {
     try {
-        const userIdFromClient = req.body.UserID || req.user.UserID;
+        const userIdFromClient = req.user.Role === "User" ? req.user.UserID : req.body.UserID
 
-        if (req.user.UserID !== userIdFromClient && req.user.Role !== "Admin") {
-            return res.status(403).json({
-                message: "Bạn không có quyền tạo đơn hàng cho người khác!",
-                your_UserID: req.user.UserID
-            });
-        }
-
-        const { ProductID, Quantity } = req.body;
+        const { ProductID, Quantity, ShippingAddress, Phone } = req.body;
         const UserID = userIdFromClient;
 
         const checkUser = await sql.query`SELECT * FROM Users WHERE UserID = ${UserID}`;
         if (checkUser.recordset.length === 0) {
             return res.status(404).send("Không có User để thêm Orders");
+        }
+
+        if ((ShippingAddress && !Phone) || (!ShippingAddress && Phone)) {
+            return res.status(400).send("Phải nhập cả địa chỉ và số điện thoại hoặc không nhập cả hai để lấy mặc định.");
+        }
+        let finalAddress = ShippingAddress;
+        let finalPhone = Phone;
+        const userInfo = checkUser.recordset[0];
+        if (!Phone && !ShippingAddress) {
+            finalAddress = userInfo.Address || "";  
+            finalPhone = userInfo.Phone || "";
+            if (!finalAddress || !finalPhone) {
+                return res.status(400).send("Vui lòng cung cấp địa chỉ và số điện thoại hoặc cập nhật trong hồ sơ người dùng.");
+            }
+        }
+        if (!userInfo.Phone || !userInfo.Address) {
+            await sql.query`Update Users 
+                SET Phone = ${Phone},
+                    Address = ${ShippingAddress}
+                WHERE UserID = ${UserID}`
         }
 
         const resultProduct = await sql.query`SELECT * FROM Products WHERE ProductID = ${ProductID}`;
@@ -90,9 +123,10 @@ async function addOrders(req, res) {
         }
 
         await sql.query`
-            INSERT INTO Orders (UserID, ProductID, Quantity)
-            VALUES(${UserID}, ${ProductID}, ${Quantity})
+            INSERT INTO Orders (UserID, ProductID, Quantity, ShippingAddress, Phone, Status)
+            VALUES(${UserID}, ${ProductID}, ${Quantity}, ${finalAddress}, ${finalPhone}, 'Pending')
         `;
+
         //cập nhật sl kho
         await sql.query`
           UPDATE Products
@@ -134,14 +168,12 @@ async function updateOrder(req, res) {
                 message: "Bạn không có quyền sửa đơn hàng này!",
             });
         }
-
+        
+        if (order.status !== "Pending") {
+            let statusMessage = getStatusOrder(order.Status);
+            return res.status(403).json({ message: `không thể sửa đơn hàng! ${statusMessage}` })
+        }
         if (ProductID !== order.ProductID) {
-            await sql.query`
-                UPDATE Products
-                SET Stock = Stock + ${order.Quantity}
-                WHERE ProductID = ${order.ProductID}
-            `;
-
             const productNew = await sql.query`SELECT * FROM Products WHERE ProductID = ${ProductID}`;
             if (productNew.recordset.length === 0) {
                 return res.status(404).send("Không tìm thấy sản phẩm mới");
@@ -157,6 +189,13 @@ async function updateOrder(req, res) {
                 SET Stock = Stock - ${Quantity}
                 WHERE ProductID = ${ProductID}
             `;
+
+            await sql.query`
+                UPDATE Products
+                SET Stock = Stock + ${order.Quantity}
+                WHERE ProductID = ${order.ProductID}
+            `;
+
         } else {
             const infoProduct = await sql.query`SELECT * FROM Products WHERE ProductID = ${ProductID}`;
             const stockProduct = infoProduct.recordset[0].Stock;
@@ -193,6 +232,15 @@ async function updateOrder(req, res) {
         res.status(500).send("Lỗi: " + err.message);
     }
 }
+function getStatusOrder(status) {
+    switch (status) {
+        case 'Pending': return "Đơn hàng đang chờ xác nhận.";
+        case 'Confirmed': return "Đơn hàng đã được xác nhận.";
+        case 'Shipped': return "Đơn hàng đang được vận chuyển.";
+        case 'Delivered': return "Đơn hàng đã giao thành công.";
+        default: return "Trạng thái đơn hàng không hợp lệ.";
+    }
+}
 
 async function deleteOrder(req, res) {
     try {
@@ -208,6 +256,11 @@ async function deleteOrder(req, res) {
             return res.status(403).json({
                 message: "Bạn không có quyền xóa đơn hàng này!",
             });
+        }
+                      
+        if (order.Status !== 'Pending') {
+            let statusMessage = getStatusOrder(order.Status);
+            return res.status(403).json({ message: `Bạn không thể xóa đơn hàng! ${statusMessage}` })
         }
 
         await sql.query`DELETE FROM Orders WHERE OrderID = ${OrderID}`;
